LLVMLINUX: kbuild: Add better clang cross build support
Add cross target to CC if using clang. Also add custom gcc toolchain path for
fallback gcc tools.

COMPILER is previously set to "clang" if CC=clang was set from the make command
line.  So -target and -gcc-toolchain can be added to CC, since we already know
that it is set.

Clang will fallback to using things like ld, as, and libgcc if (respectively)
one of the llvm linkers isn't available, the integrated assembler is turned
off, or an appropriately cross-compiled version of compiler-rt isn't available.
To this end, you can specify the path to this fallback gcc toolchain with
GCC_TOOLCHAIN.

Signed-off-by: Behan Webster <behanw@converseincode.com>
Reviewed-by: Jan-Simon Möller <dl9pf@gmx.de>
Reviewed-by: Mark Charlebois <charlebm@gmail.com>
Signed-off-by: Greg Hackmann <ghackmann@google.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: AyushR1 <ayushrathore12501@gmail.com>
Signed-off-by: Joshua Primero <jprimero1555@gmail.com>
Change-Id: I64890a1bdbe9398e9070617e2c5d32345cf20dd6
Signed-off-by: Joshua Primero <jprimero1555@gmail.com>